refactor(menu): extract shared item modal
